backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
python-dotenv==1.0.1
python-multipart==0.0.18
pydantic==2.9.2
pydantic-settings
google-genai>=1.0.0
requests==2.32.3
email-validator
pillow
numpy
torch
torchvision
